Ingresse AWS - Project initial commit
This project is a tool to execute actions inside aws ecosystem.
The fist needs is start, stop instances and also check your status.
Schudule was a submodule defined to this interation. The necessity
of this is to avoid instances to be on certains hours that it is
not being used.

Readme contains info about setup and configuration.